Add files via upload
This project aims to examine gender equality in the workplace through data analysis and visualization. Developed in a Jupyter Notebook using Python, it analyzes metrics such as the ratio of male to female employees, pay gaps, and representation in managerial positions. The findings are intended to highlight gender disparities in professional environments and raise awareness about workplace equality. The project utilizes libraries such as Pandas, NumPy, Matplotlib, and Seaborn, with results supported by various visualizations.